not_if uaa

+++ dev_setup/cookbooks/uaa/recipes/default.rb
@@ -24,6 +24,7 @@ bash "Build and Deploy UAA" do
     rm -Rf #{node[:tomcat][:base]}/webapps/ROOT
     cp -f #{node[:uaa][:path]}/uaa/target/cloudfoundry-identity-uaa-*.war #{node[:tomcat][:base]}/webapps/ROOT.war
   EOH
+  not_if "ls -la #{File.join(node[:uaa][:path], 'uaa', 'target')} | grep `grep -m 1 \<version\> #{File.join(node[:uaa][:path], "pom.xml")} | sed -n -e 's/<version>\([0-9]\.[0-9]\.[0-9]\)<\/version>/\1/p'`.war"
 end
 
 cf_bundle_install(File.expand_path("uaa", node["cloudfoundry"]["path"]))
